moved LayerSpec to spec package
