+++ src/main/scala/hohserg/dimensional/layers/gui/add/dimension/GuiAddLayer.scala
@@ -3,7 +3,7 @@ package hohserg.dimensional.layers.gui.add.dimension
 import hohserg.dimensional.layers.gui.GuiSelectDimension
 import hohserg.dimensional.layers.gui.GuiSelectDimension.DrawableDim
 import hohserg.dimensional.layers.gui.preset.GuiSetupDimensionalLayersPreset
-import hohserg.dimensional.layers.preset.DimensionLayerSpec
+import hohserg.dimensional.layers.preset.spec.DimensionLayerSpec
 import net.minecraftforge.fml.relauncher.{Side, SideOnly}
 
 @SideOnly(Side.CLIENT)
