+++ src/main/scala/hohserg/dimensional/layers/gui/preset/list/GuiCubicWorldTypeLayerEntry.scala
@@ -4,7 +4,7 @@ import hohserg.dimensional.layers.gui.GuiBase
 import hohserg.dimensional.layers.gui.IconUtils._
 import hohserg.dimensional.layers.gui.preset.GuiSetupDimensionalLayersPreset
 import hohserg.dimensional.layers.gui.settings._
-import hohserg.dimensional.layers.preset.CubicWorldTypeLayerSpec
+import hohserg.dimensional.layers.preset.spec.CubicWorldTypeLayerSpec
 import net.minecraftforge.fml.relauncher.{Side, SideOnly}
 
 @SideOnly(Side.CLIENT)
